requirements.txt
asyncio
openai-whisper
setuptools-rust
colorama
tqdm
pydub
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
git+https://github.com/m-bain/whisperx.git
langchain[all]
yt_dlp
faster-whisper